Some fun with exceptions
Possible solution to the challenge:

Write a function that given a list will return True if all the integers in the
list are in a numeric range with no missing numbers. If not, raise a custom
exception.

Given [1, 2, 3, 4] returns True
Given [2, 3, 4] returns True
Given [4, 6, 5, 2, 3] returns True
Given [1, 4, 7] raises and Exception

Bonus: write unit tests, neatly handle case when something that is not a list
is passed to your function, if the list passed contains something other than an
integer raise a custom exception