An update of my mini-project.
I removed the random_encounters function, added a list and a function
that will set a 'lose' conditional (the program is planting evidence against the player as the story progresses) , removed the Death class and turned it into a function, wrote more of the story (and included the first instance of the new death function) and began to comment where it seemed appropriate. Next step: write many more scenes for the story!